test: move test to right test class
